+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 1.x branch
 ## 1.1 branch
+### 1.1.9
+* `file_exists_or_fail()`, `is_dir_or_fail()`, `is_readable_or_fail()` and
+    `is_writable_or_fail()` functions now have `$message` the second argument.
+
 ### 1.1.8
 * `create_tmp_file()` and `safe_create_tmp_file()` methods now accept a 
     directory as a second argument and a prefix as the third argument.